final-pjt-back/requirements.txt
asgiref==3.4.1
boto==2.49.0
boto3==1.20.12
botocore==1.23.12
certifi==2021.10.8
charset-normalizer==2.0.7
cleanup==1.0.3
dj-database-url==0.5.0
Django==3.2.6
django-appconf==1.0.5
django-cors-headers==3.10.0
django-dump-load-utf8==0.0.4
django-imagekit==4.1.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
gunicorn==20.1.0
huepy==1.2.1
idna==3.3
jmespath==0.10.0
pilkit==2.0
Pillow==8.4.0
psycopg2==2.9.2
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
whitenoise==5.3.0
